# src/auth_server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OAuth2 Server with OTP Authentication",
    "description": "A production-ready OAuth2 authorization server with TOTP authentication",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
    "/register_user": {
      "post": {
        "summary": "Register a new user",
        "tags": ["User Management"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {"type": "string", "example": "testUser"},
                  "password": {"type": "string", "example": "testPassword123"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "otp_uri": {"type": "string"},
                    "qr_code_url": {"type": "string"}
                  }
                }
              }
            }
          },
          "400": {"description": "Bad request"}
        }
      }
    },
    "/qr_code/{username}": {
      "get": {
        "summary": "Get QR code image",
        "tags": ["User Management"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "QR code image",
            "content": {"image/png": {}}
          },
          "404": {"description": "User not found"}
        }
      }
    },
    "/verify_otp": {
      "post": {
        "summary": "Verify OTP code",
        "tags": ["User Management"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "code"],
                "properties": {
                  "username": {"type": "string", "example": "testUser"},
                  "code": {"type": "string", "example": "123456"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {"type": "boolean"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/create_client": {
      "post": {
        "summary": "Create OAuth2 client",
        "tags": ["Client Management"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["client_name", "redirect_uri"],
                "properties": {
                  "client_name": {"type": "string", "example": "MyApp"},
                  "redirect_uri": {"type": "string", "example": "https://127.0.0.1:5000/callback"},
                  "grant_types": {
                    "type": "array",
                    "items": {"type": "string"},
                    "example": ["client_credentials", "authorization_code", "password"]
                  },
                  "scope": {"type": "string", "example": "profile email"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Client created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "client_id": {"type": "string"},
                    "client_secret": {"type": "string"}
                  }
                }
              }
            }
          },
          "400": {"description": "Bad request"}
        }
      }
    },
    "/token": {
      "post": {
        "summary": "Get access token",
        "tags": ["OAuth2"],
        "security": [{"basicAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": ["grant_type"],
                "properties": {
                  "grant_type": {
                    "type": "string",
                    "enum": ["password", "client_credentials", "refresh_token"],
                    "example": "password"
                  },
                  "username": {"type": "string", "example": "testUser"},
                  "password": {"type": "string", "example": "testPassword123"},
                  "scope": {"type": "string", "example": "profile"},
                  "refresh_token": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token issued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {"type": "string"},
                    "token_type": {"type": "string"},
                    "expires_in": {"type": "integer"},
                    "refresh_token": {"type": "string"},
                    "scope": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/profile": {
      "get": {
        "summary": "Get user profile",
        "tags": ["Protected Resources"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Profile information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {"type": "string"},
                    "username": {"type": "string"},
                    "otp_verified": {"type": "boolean"},
                    "client_id": {"type": "string"},
                    "message": {"type": "string"}
                  }
                }
              }
            }
          },
          "401": {"description": "Unauthorized"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}